docs: Clean README and fix contract test arithmetic overflow
README Updates:
- Remove external service references (testnet.citrate.ai, docs.citrate.ai, etc.)
- Remove explorer/docs deployment instructions per user request
- Update SDK examples to use localhost instead of external URLs
- Remove Vercel deployment badge and instructions
- Update community links to point to GitHub repo

Contract Fix:
- Fix testAddReview arithmetic overflow (uint8 * 100 exceeded max 255)
- Cast rating to uint256 before multiplication

Sprint Planning:
- Add Sprint 11 plan for GUI Final Delivery & Documentation
- Update CURRENT_SPRINT.md to reflect active sprint

Test Status:
- Rust: 314/315 passing (1 pre-existing async runtime issue)
- Contracts: 43/43 passing (all tests now pass)

+++ citrate/contracts/test/ModelMarketplace.t.sol
@@ -314,7 +314,7 @@ contract ModelMarketplaceTest is Test {
         ModelMarketplace.ModelListing memory listing = marketplace.getListing(testModelId);
         assertEq(listing.reviewCount, 1);
         assertEq(listing.totalRating, rating);
-        assertEq(listing.averageRating, rating * 100); // Stored as percentage
+        assertEq(listing.averageRating, uint256(rating) * 100); // Stored as percentage
     }
 
     function testAddReviewUnverified() public {
